requirements.txt
# LLM & AI Agent Frameworks
langchain==0.1.12
openai==1.14.1
llama-index  # Automatically installs the latest version

# Backend (API)
fastapi==0.109.2
uvicorn==0.27.0.post1

# Database
sqlalchemy==2.0.27
sqlite-utils==3.35




# Environment Variables
python-dotenv==1.0.1


# Logging
loguru==0.7.2


# Task Scheduling
APScheduler==3.10.4


# CLI Integration 
typer==0.9.0


# Streamlit (Frontend)
streamlit==1.31.1